Revert "frontend rebuild. Skeleton landing page up"
This reverts commit fd89a337724f1ac9afce07a6de966e78fbf37cb8.
