--- client/src/components/About.js
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const About = () => (
-	<div>
-		<p>About</p>
-		<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
-	</div>
-)
-
-export default About
