--- client/src/components/Archive/LandingPage.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const LandingPage = () => (
-	<div>
-		<p>This is the landing page!</p>
-	</div>
-)
-
-export default LandingPage
