--- client/src/components/Contact.js
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const Contact = () => (
-	<div>
-		<p>Contact Us!</p>
-		<p>info@codeforboulder.org</p>
-	</div>
-)
-
-export default Contact
