--- client/src/components/DonationMap.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const DonationMap = () => (
-	<div>
-		<p>This is the map section</p>
-	</div>
-)
-
-export default DonationMap
