--- client/src/components/FilterBoxes.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const FilterBoxes = () => (
-	<div>
-		<p>This is the box filter section</p>
-	</div>
-)
-
-export default FilterBoxes
