--- client/src/components/ItemSearchBox.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const ItemSearchBox = () => (
-	<div>
-		<p>This is the item search box section</p>
-	</div>
-)
-
-export default ItemSearchBox
