+++ client/src/components/LandingPage.js
@@ -1,38 +1,9 @@
 import React from 'react'
-import Header from './Header'
-import About from './About'
-import LocationSearch from './LocationSearch'
-import FilterBoxes from './FilterBoxes'
-import ItemSearchBox from './ItemSearchBox'
-import DonationMap from './DonationMap'
-import Contact from './Contact'
-
 
 const LandingPage = () => (
 	<div>
-		<div>
-			<Header/>
-		</div>
-		<div>
-			<About/>
-		</div>
-		<div>
-			<LocationSearch/>
-		</div>
-		<div>
-			<FilterBoxes/>
-		</div>
-		<div>
-			<ItemSearchBox/>
-		</div>
-		<div>
-			<DonationMap/>
-		</div>
-		<div>
-			<Contact/>
-		</div>
+		<p>This is the landing page!</p>
 	</div>
 )
 
 export default LandingPage
-
